RatingPredictionModel (wip)

+++ src/main/java/core/training/TrainingPerPlayer.java
@@ -77,7 +77,7 @@ public class TrainingPerPlayer  {
 		return experienceSub;
 	}
 
-    public float calcSubskillIncrement(int skill, float skillValueBeforeTraining, HODateTime date) {
+    public float calcSubskillIncrement(int skill, double skillValueBeforeTraining, HODateTime date) {
 
 		int skillValue = (int) skillValueBeforeTraining;
 		float ret = 0;
